Task: Se validaron los formularios del proyecto y tambien se actualizo las rutas de acceso

+++ TRABAJO_FINAL_GRUPO_STROKES/src/main/java/ar/edu/unju/fi/tpfinal/service/imp/EmployeeServiceImp.java
@@ -57,7 +57,7 @@ public class EmployeeServiceImp implements IEmployeeService {
 	}
 
 	@Override
-	public Employee searchEmployee(int numero, String email) {
+	public Employee searchEmployee(Long numero, String email) {
 		return employeeRepository.findByEmployeeNumberAndEmailGreaterThanEqual(numero, email);
 	}
